# spec/openapi/iroute.v1.yaml
openapi: 3.1.0
info:
  title: iRoute Runtime API
  version: 1.0.0
  description: Task-aware execution, event streaming, reusable artifacts, and privacy-safe observability API.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - {}
paths:
  /v1/executions:
    post:
      operationId: createExecution
      tags: [Executions]
      summary: Persist and submit one typed task for execution.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/PermissionScopes'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '202':
          description: The execution was persisted and is queued, resolved by the no-model fast path, or paused in `WaitingForApproval`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionSnapshot'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/executions/{executionId}:
    get:
      operationId: getExecution
      tags: [Executions]
      summary: Get the current execution snapshot.
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Current execution state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionSnapshot'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /v1/executions/{executionId}/events:
    get:
      operationId: streamExecutionEvents
      tags: [Executions]
      summary: Stream ordered execution events using Server-Sent Events.
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
        - $ref: '#/components/parameters/TenantId'
        - name: after
          in: query
          description: Return events after this sequence number.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: Last-Event-ID
          in: header
          description: SSE reconnection cursor used when `after` is omitted.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Each SSE `data` field contains one serialized ExecutionEvent.
          x-sse-frame:
            id: TaskEvent.sequence
            event: TaskEvent.type
            data: TaskEvent JSON
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/TaskEvent'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /v1/executions/{executionId}/cancel:
    post:
      operationId: cancelExecution
      tags: [Executions]
      summary: Request cancellation of a non-terminal execution.
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
        - $ref: '#/components/parameters/TenantId'
      responses:
        '202':
          description: Cancellation was recorded and signalled to the local executor.
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /v1/executions/{executionId}/approvals:
    post:
      operationId: submitExecutionApproval
      tags: [Executions]
      summary: Approve or deny a pending external action and resume the durable workflow when approved.
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/PermissionScopes'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecision'
      responses:
        '200':
          description: The approval decision and current execution snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/artifacts/{artifactId}:
    get:
      operationId: getArtifact
      tags: [Artifacts]
      summary: Get a materialized artifact and its provenance.
      parameters:
        - $ref: '#/components/parameters/ArtifactId'
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Artifact content and metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactSnapshot'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
  /v1/observability/executions/{executionId}:
    get:
      operationId: getExecutionTimeline
      tags: [Observability]
      summary: Inspect one redacted execution timeline and its trace correlation identifier.
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Ordered and bounded execution timeline. Event data is metadata-only by default; recursively redacted fields are an operator opt-in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionTimeline'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Required when `Identity:Mode` is `Jwt`.
  parameters:
    ExecutionId:
      name: executionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: Development-only tenant scope. JWT mode ignores this header and uses the configured tenant claim.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    ActorId:
      name: X-Actor-Id
      in: header
      required: false
      description: Development-only actor scope. JWT mode ignores this header and uses the configured actor claim.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    PermissionScopes:
      name: X-Permission-Scopes
      in: header
      required: false
      description: Development-only space- or comma-separated permission scopes. JWT mode ignores this header and uses the configured permission claim.
      schema:
        type: string
        maxLength: 4096
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Tenant-scoped idempotency key. A body value must match this header.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  responses:
    Problem:
      description: RFC 9457 problem details with a stable iRoute error code.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/HttpProblem'
  schemas:
    TaskRequest:
      type: object
      additionalProperties: false
      required: [taskType, input]
      properties:
        taskType:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '^[a-z][a-z0-9_.-]+$'
        input: {}
        projectId:
          type: [string, 'null']
          maxLength: 200
        idempotencyKey:
          type: [string, 'null']
          maxLength: 200
        constraints:
          oneOf:
            - $ref: '#/components/schemas/TaskConstraints'
            - type: 'null'
        metadata:
          type: [object, 'null']
          description: Optional routing metadata. The reserved `artifactKey` value selects an artifact lineage and is limited to 200 characters.
          additionalProperties:
            type: string
          maxProperties: 32
        tenantId:
          type: [string, 'null']
          maxLength: 200
          description: Development scope hint. In JWT mode, a supplied value must match the authenticated claim.
        actorId:
          type: [string, 'null']
          maxLength: 200
          description: Development scope hint. In JWT mode, a supplied value must match the authenticated claim.
        permissionScopes:
          type: [array, 'null']
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 200
          description: Server-derived authorization context. Request-body values are ignored and replaced by authenticated scopes.
    TaskConstraints:
      type: object
      additionalProperties: false
      properties:
        maxInputTokens: { type: [integer, 'null'], minimum: 1, maximum: 2000000 }
        maxOutputTokens: { type: [integer, 'null'], minimum: 1, maximum: 200000 }
        maxCost: { type: [number, 'null'], minimum: 0 }
        deadlineMilliseconds: { type: [integer, 'null'], minimum: 1, maximum: 900000 }
        minimumQuality: { type: [number, 'null'], minimum: 0, maximum: 1 }
        requireEvidence: { type: boolean, default: false }
        allowExternalWrites: { type: boolean, default: false }
        maxModelCalls: { type: [integer, 'null'], minimum: 0, maximum: 16 }
        maxToolCalls: { type: [integer, 'null'], minimum: 0, maximum: 64 }
        maxParallelCalls: { type: [integer, 'null'], minimum: 1, maximum: 32 }
        maxTaskDepth: { type: [integer, 'null'], minimum: 1, maximum: 16 }
    ExecutionSnapshot:
      type: object
      additionalProperties: false
      required: [executionId, taskType, status, createdAt, updatedAt, tenantId, actorId]
      properties:
        executionId: { type: string, format: uuid }
        taskType: { type: string }
        status: { $ref: '#/components/schemas/ExecutionStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        outcome:
          oneOf:
            - $ref: '#/components/schemas/TaskOutcome'
            - type: 'null'
        error:
          oneOf:
            - $ref: '#/components/schemas/ExecutionProblem'
            - type: 'null'
        tenantId: { type: string }
        actorId: { type: string }
        projectId: { type: [string, 'null'] }
        taskDefinitionVersion: { type: [integer, 'null'], minimum: 1 }
        cancellationRequestedAt: { type: [string, 'null'], format: date-time }
    ApprovalDecision:
      type: object
      additionalProperties: false
      required: [actionId, approved]
      properties:
        actionId: { type: string, minLength: 1, maxLength: 200 }
        approved: { type: boolean }
        reason: { type: [string, 'null'], maxLength: 2000 }
    ApprovalSnapshot:
      type: object
      additionalProperties: false
      required: [executionId, actionId, status, capability, sideEffectClass, requiredPermissionScopes, requestedByActorId, inputReference, idempotencyReference, createdAt]
      properties:
        executionId: { type: string, format: uuid }
        actionId: { type: string }
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        capability: { type: string }
        sideEffectClass: { $ref: '#/components/schemas/SideEffectClass' }
        requiredPermissionScopes:
          type: array
          uniqueItems: true
          items: { type: string }
        requestedByActorId: { type: string }
        decidedByActorId: { type: [string, 'null'] }
        inputReference: { type: string }
        idempotencyReference: { type: string }
        createdAt: { type: string, format: date-time }
        decidedAt: { type: [string, 'null'], format: date-time }
        reason: { type: [string, 'null'] }
    ApprovalResult:
      type: object
      additionalProperties: false
      required: [approval, execution]
      properties:
        approval: { $ref: '#/components/schemas/ApprovalSnapshot' }
        execution: { $ref: '#/components/schemas/ExecutionSnapshot' }
    ApprovalStatus:
      type: string
      enum: [Pending, Approved, Denied]
    SideEffectClass:
      type: string
      enum: [None, ReadOnly, ReversibleWrite, IrreversibleWrite]
    ExecutionStatus:
      type: string
      enum:
        - Accepted
        - Resolving
        - Planning
        - Queued
        - WaitingForApproval
        - Running
        - Validating
        - Materializing
        - Compensating
        - Succeeded
        - Failed
        - Cancelled
        - TimedOut
    TaskOutcome:
      type: object
      additionalProperties: false
      required: [output, resolutionLevel, confidence, evidence, usage, artifacts]
      properties:
        output: {}
        resolutionLevel:
          $ref: '#/components/schemas/ResolutionLevel'
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidence:
          type: array
          items: { $ref: '#/components/schemas/EvidenceReference' }
        usage: { $ref: '#/components/schemas/UsageSummary' }
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/ArtifactReference' }
        validation:
          oneOf:
            - $ref: '#/components/schemas/ValidationSummary'
            - type: 'null'
        context:
          oneOf:
            - $ref: '#/components/schemas/ContextManifest'
            - type: 'null'
        routing:
          oneOf:
            - $ref: '#/components/schemas/RoutingDecision'
            - type: 'null'
    ResolutionLevel:
      type: string
      enum:
        - ExactArtifact
        - StructuredState
        - SemanticMemory
        - DeterministicCapability
        - SmallModel
        - StrongModel
        - VerifiedOrHuman
    EvidenceReference:
      type: object
      additionalProperties: false
      required: [kind, reference]
      properties:
        kind: { type: string }
        reference: { type: string }
        contentHash: { type: [string, 'null'] }
        observedAt: { type: [string, 'null'], format: date-time }
    DependencyReference:
      type: object
      additionalProperties: false
      required: [kind, reference]
      properties:
        kind: { type: string, minLength: 1, maxLength: 120 }
        reference: { type: string, minLength: 1, maxLength: 500 }
        contentHash: { type: [string, 'null'], maxLength: 256 }
    ArtifactReference:
      type: object
      additionalProperties: false
      required: [artifactId, artifactType, version, contentHash]
      properties:
        artifactId: { type: string, format: uuid }
        artifactType: { type: string }
        version: { type: integer, minimum: 1 }
        contentHash: { type: string }
    ArtifactSnapshot:
      type: object
      additionalProperties: false
      required:
        - artifact
        - tenantId
        - taskType
        - taskDefinitionVersion
        - content
        - evidence
        - createdAt
        - isActive
      properties:
        artifact: { $ref: '#/components/schemas/ArtifactReference' }
        tenantId: { type: string }
        projectId: { type: [string, 'null'] }
        taskType: { type: string }
        taskDefinitionVersion: { type: integer, minimum: 1 }
        content: {}
        evidence:
          type: array
          items: { $ref: '#/components/schemas/EvidenceReference' }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: [string, 'null'], format: date-time }
        isActive: { type: boolean }
        logicalKey: { type: [string, 'null'], minLength: 1, maxLength: 200 }
        lifecycleStatus: { $ref: '#/components/schemas/ArtifactLifecycleStatus' }
        supersedesArtifactId: { type: [string, 'null'], format: uuid }
        supersededByArtifactId: { type: [string, 'null'], format: uuid }
        dependencies:
          type: [array, 'null']
          items: { $ref: '#/components/schemas/DependencyReference' }
        invalidatedAt: { type: [string, 'null'], format: date-time }
        invalidationReason: { type: [string, 'null'], maxLength: 2000 }
    ArtifactLifecycleStatus:
      type: string
      enum: [Active, Superseded, Invalidated]
...
    ExecutionTimeline:
      type: object
      additionalProperties: false
      required: [executionId, traceId, taskType, taskDefinitionVersion, policyVersion, status, startedAt, updatedAt, durationMilliseconds, actorReference, projectReference, payloadMode, truncated, events]
      properties:
        executionId: { type: string, format: uuid }
        traceId: { type: string, minLength: 1, maxLength: 64 }
        taskType: { type: string }
        taskDefinitionVersion: { type: [integer, 'null'], minimum: 1 }
        policyVersion: { type: string }
        status: { $ref: '#/components/schemas/ExecutionStatus' }
        startedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        durationMilliseconds: { type: integer, format: int64, minimum: 0 }
        actorReference: { type: string, pattern: '^sha256:[a-f0-9]{16}$' }
        projectReference: { type: [string, 'null'], pattern: '^sha256:[a-f0-9]{16}$' }
        payloadMode: { $ref: '#/components/schemas/ObservabilityPayloadMode' }
        truncated: { type: boolean }
        events:
          type: array
          items: { $ref: '#/components/schemas/ExecutionTimelineEntry' }
    ExecutionTimelineEntry:
      type: object
      additionalProperties: false
      required: [sequence, type, occurredAt, offsetMilliseconds, data]
      properties:
        sequence: { type: integer, format: int64, minimum: 1 }
        type: { type: string }
        occurredAt: { type: string, format: date-time }
        offsetMilliseconds: { type: integer, format: int64, minimum: 0 }
        data:
          type: object
          additionalProperties: true
    ObservabilityPayloadMode:
      type: string
      enum: [MetadataOnly, Redacted]
...
    TaskEvent:
      type: object
      additionalProperties: false
      required: [sequence, executionId, type, occurredAt, data]
      properties:
        sequence: { type: integer, format: int64, minimum: 1 }
        executionId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/TaskEventType' }
        occurredAt: { type: string, format: date-time }
        data:
          type: object
          additionalProperties: true
...
    TaskEventType:
      type: string
      description: Additive event types may be introduced in v1; consumers must ignore unknown types.
      enum:
        - execution.created
        - execution.status_changed
        - execution.queued
        - execution.lease_claimed
        - execution.lease_renewed
        - execution.lease_released
        - resolution.considered
        - routing.decided
        - routing.escalated
        - plan.validated
        - policy.evaluated
        - capability.denied
        - capability.started
        - capability.completed
        - capability.failed
        - approval.required
        - approval.decided
        - workflow.checkpointed
        - workflow.resumed
        - step.started
        - step.completed
        - step.retry_scheduled
        - step.failed
        - external_action.started
        - external_action.completed
        - external_action.reused
        - external_action.failed
        - context.compiled
        - gateway.started
        - gateway.streamed
        - gateway.completed
        - gateway.failed
        - validation.completed
        - artifact.materialized
        - artifact.superseded
        - artifact.invalidated
        - memory.materialized
        - memory.superseded
        - memory.invalidated
        - execution.cancellation_requested
        - execution.completed
        - execution.failed
...
    ExecutionProblem:
      type: object
      additionalProperties: false
      required: [code, title, detail, retryable]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        title: { type: string }
        detail: { type: string }
        retryable: { type: boolean }
        metadata:
          type: [object, 'null']
          additionalProperties: { type: string }
    HttpProblem:
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 100, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { $ref: '#/components/schemas/ErrorCode' }